fixed random bug => blavklisted pixel ninja frmo using c.say and c.tsay commands.

+++ commands/tsay.js
@@ -3,7 +3,7 @@ const Discord = require("discord.js");
 module.exports.run = async (bot, message, args) => {
    // if(message.author.id !== "501710994293129216")return message.reply("Only the owner**s** cam use this commands!")
 
-
+   if(message.author.id === "437255943181565962") return message.reply(`Due to your recent actions, you have been blacklisted from using this command! \n \b \b \b \b`)
       const sayMessage = args.join(" ");
       if(!sayMessage) return message.reply("Make sure you tell me what you want me to say...")
       message.delete().catch();
